Stop trying to be beautiful and simply print commands as expected from make
We used to hide the commands and print ansi-based chrome to show what
was happening. Although that was somewhat helpful to see what was going
on as an end-user, it also obscured commands that were run. Since I'm
reworking a project that uses this build-system to move to a simpler one,
I want to make operations as transparent as they can.

You can always pass `-s` to ysr / make to silence the echoing of recipes.