Use GNU Make 3.81's secondary expansion to automatically create directories
We now make use of the secondary expansion feature to expand functions /
variables in pre-requisites, which allows us to replace calls to
`require-directory` in each build rule. Hopefully this makes directory
creation faster too, in addition to making the rules smaller.

The syntax to use to automatically create the destination's directory
is:

```Makefile
%.o: %.c $$(@D)/.dir
```

It makes use of the automatic variable $(@D) which expands to the
directory of the target, prefixed with a double $ to allow it to
take effect during secondary expansion.